clippy: fix warnings in tests

+++ tests/bootstrap04.rs
@@ -14,8 +14,8 @@ fn create_new_xlsx_file_1(filename: &str) -> Result<(), XlsxError> {
     let mut workbook = Workbook::new();
     let worksheet = workbook.add_worksheet();
 
-    worksheet.write_number(0, 0, 1 as u32)?;
-    worksheet.write_number(1, 1, 2 as f32)?;
+    worksheet.write_number(0, 0, 1)?;
+    worksheet.write_number(1, 1, 2)?;
     worksheet.write_number(2, 2, 3)?;
 
     workbook.save(filename)?;
@@ -28,8 +28,8 @@ fn create_new_xlsx_file_2(filename: &str) -> Result<(), XlsxError> {
     let mut workbook = Workbook::new();
     let worksheet = workbook.add_worksheet();
 
-    worksheet.write(0, 0, 1 as u32)?;
-    worksheet.write(1, 1, 2 as f32)?;
+    worksheet.write(0, 0, 1)?;
+    worksheet.write(1, 1, 2)?;
     worksheet.write(2, 2, 3)?;
 
     workbook.save(filename)?;
